Data/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
Deprecated==1.2.13
Django==3.2.12
django-cors-headers==3.13.0
django-redis==5.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
joblib==1.2.0
mysqlclient==2.1.1
numpy==1.23.3
packaging==21.3
pandas==1.4.4
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
redis==4.3.4
scikit-learn==1.1.2
scikit-surprise==1.1.2
scipy==1.9.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
wrapt==1.14.1
